fix errors tested on vehicle

+++ cmd_actuators_mux/src/cmd_actuator_mux_nodelet.cpp
@@ -397,9 +397,10 @@ void CmdActuatorMuxNodelet::reloadConfiguration(cmd_actuators_mux::reloadConfig&
 
   // check that the both arrays have the same length 
   ROS_ASSERT_MSG(cmd_set_rpm_subs.size()==(cmd_fin_angles_subs.size()),
-                   "Different amount of inputs between SetRPM Subscriber (%ld) and "
-                   "Fin_Angles Subscribers (%ld) in YAML file:",
-                   cmd_set_rpm_subs.size(), cmd_fin_angles_subs.size());
+                   "Different amount of inputs between SetRPM Subscriber (%d) and "
+                   "Fin_Angles Subscribers (%d) in YAML file:",
+                   static_cast<int>(cmd_set_rpm_subs.size()),
+                   static_cast<int>(cmd_fin_angles_subs.size()));
 
   // check that the subscribers input have the same name in YAML
   for (int i = 0; i < cmd_set_rpm_subs.size(); i++)
